Make event processor implementation to use message queue
